Move migration functions/code into dedicated file
to not blow / overload the `app.go`
